Refactor tool tips: Sidebar and RuleCalculation now use the same tooltip icon component   - Add circle icon box shadow styling back in   - Create a prop for different icon sizes with small as default #471
